requirements.txt
PySide6>=6.6.0
requests>=2.31.0
Pillow>=10.0.0
rembg>=2.0.50
